Refactor authentication and utility imports; streamline CourseSelector component
