+++ src/components/common/utils/route.ts
@@ -1,4 +1,5 @@
 //export const porturl = "https://pfz60p76-3000.brs.devtunnels.ms";
-export const porturl = import.meta.env.VITE_API_URL || "http://localhost:3000";
+const url = import.meta.env.VITE_API_URL;
+export const porturl = "http://localhost:3000";
 //export const porturl = "https://tp-dsw-backend.onrender.com";
 //esto lo pueden editar para que funcione en su entorno cualquier cosa pregunten por el grupo de whatsapp
